Clean up dead IdentityCard; drop emoji labels; document topRepos deviation
