--- web/src/components/IdentityCard.css
@@ -1,50 +1,0 @@
-.idcard {
-  max-width: var(--column-max);
-  margin: 0 auto;
-  padding: 48px var(--side-gutter) 28px;
-  color: var(--text);
-}
-
-.idcard-name {
-  font-size: 34px;
-  font-weight: 700;
-  line-height: 1.05;
-  letter-spacing: -0.01em;
-  margin: 0;
-  opacity: 0;
-  animation: idcard-in 280ms ease-out 80ms forwards;
-}
-
-.idcard-role {
-  font-size: 13px;
-  color: var(--text-muted);
-  margin-top: 10px;
-  opacity: 0;
-  animation: idcard-in 280ms ease-out 160ms forwards;
-}
-
-.idcard-profiles {
-  display: flex;
-  gap: 14px;
-  flex-wrap: wrap;
-  margin-top: 18px;
-  font-size: 12px;
-  opacity: 0;
-  animation: idcard-in 280ms ease-out 320ms forwards;
-}
-
-@keyframes idcard-in {
-  from { opacity: 0; transform: translateY(6px); }
-  to   { opacity: 1; transform: translateY(0); }
-}
-
-@media (max-width: 640px) {
-  .idcard { padding: 36px var(--side-gutter-mobile) 22px; }
-  .idcard-name { font-size: 28px; }
-}
-
-@media print {
-  .idcard-name { animation: none; opacity: 1; }
-  .idcard-role, .idcard-profiles { animation: none; opacity: 1; }
-  .idcard .caret { display: none; }
-}
